refactor util to deserialization
